Correct LSEQ/RSEQ INFO specification
LSEQ and RSEQ had an INFO specification of Number=G, which is one value
per possible genotype. This is not correct with the usage, which I
believe is 1 flanking sequence per variant (Number=1). The incorrect
specification causes errors with downstream tools, like bcftools, that
check this:
```
Error at 1:10469: wrong number of fields in LSEQ?
```
Here is a small test case:

https://s3.amazonaws.com/chapmanb/az/bcftools_merge_lseq.tar.gz